fixed drop comments function
